Rename modules and delete unused code.
