+++ estimagic/__init__.py
@@ -1,5 +1,5 @@
 __version__ = "0.0.30"
 
 
-from estimagic.optimization.optimize_new import minimize  # noqa: F401
-from estimagic.optimization.optimize_new import maximize  # noqa: F401
+from estimagic.optimization.optimize import minimize  # noqa: F401
+from estimagic.optimization.optimize import maximize  # noqa: F401
